requirements3.7.txt
absl-py==1.3.0
asn1crypto==0.24.0
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2022.12.7
charset-normalizer==2.0.12
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
distlib==0.3.6
dnspython==2.2.1
entrypoints==0.3
enum34==1.1.6
et-xmlfile==1.1.0
filelock==3.4.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==3.7.0
idna==2.6
ipython-genutils==0.2.0
jedi==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyrings.alt==3.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
numpy==1.17.3
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
opt-einsum==3.3.0
pandas==1.3.5
pandas-stubs==1.2.0.62
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
platformdirs==2.4.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
pycairo==1.23.0
pycparser==2.21
pycrypto==2.6.1
Pygments==2.4.2
PyGObject==3.42.2
pymongo==4.3.3
pyparsing==2.4.2
pyrsistent==0.15.4
pysbd==0.3.4
python-dateutil==2.8.0
pytz==2022.7
pyxdg==0.25
pyzmq==18.1.0
requests==2.27.1
requests-oauthlib==1.3.1
Send2Trash==1.5.0
six==1.11.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.64.1
traitlets==4.3.3
tweepy==4.12.1
typing_extensions==4.1.1
urllib3==1.26.13
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
zipp==3.6.0
